feat: extract options and add more and others default
